Apply change to the main activity
