--- android/app/src/main/java/com/epfl/dedis/hbt/MainActivity.kt
@@ -1,28 +1,0 @@
-package com.epfl.dedis.hbt
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.viewpager.widget.ViewPager
-import com.epfl.dedis.hbt.databinding.ActivityMainBinding
-import com.epfl.dedis.hbt.ui.main.SectionsPagerAdapter
-import com.google.android.material.tabs.TabLayout
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        val sectionsPagerAdapter = SectionsPagerAdapter(this, supportFragmentManager)
-        val viewPager: ViewPager = binding.viewPager
-        viewPager.adapter = sectionsPagerAdapter
-        val tabs: TabLayout = binding.tabs
-        tabs.setupWithViewPager(viewPager)
-    }
-}
