+++ android/app/src/main/res/layout/fragment_main.xml
@@ -15,5 +15,6 @@
   <androidx.viewpager2.widget.ViewPager2
     android:id="@+id/main_pager"
     android:layout_width="match_parent"
-    android:layout_height="0dp" />
+    android:layout_height="0dp"
+    android:layout_weight="1" />
 </LinearLayout>
